feat: auto-truncate long event content + summarize_event tool
- describe_event now truncates event content > 4000 chars at word
  boundaries, preventing long-form posts from blowing up LLM context
- New summarize_event tool lets the LLM request an AI summary of
  truncated content, keeping classification context compact
- Added event ID to describe_event output so the LLM can reference it
- Updated classification prompt with truncation/summarize guidance
- Added .pi-lens/ to .gitignore
- Auto clippy fixes across several files

+++ src/db.rs
@@ -229,7 +229,7 @@ impl Database {
                         r#"INSERT INTO _sqlx_migrations (version, description, success, checksum, execution_time)
                            VALUES (?, ?, TRUE, ?, -1)"#
                     )
-                    .bind(init.version as i64)
+                    .bind(init.version)
                     .bind(&init.description)
                     .bind(init.checksum.as_ref() as &[u8])
                     .execute(pool)
